+++ sfsimodels/__about__.py
@@ -1,4 +1,4 @@
 __project__ = "sfsimodels"
 __author__ = "Maxim Millen"
-__version__ = "0.9.37"
+__version__ = "0.9.38"
 __license__ = "MIT"